Split HotsDialog.util to separate module
